export telemetry to separate module
