+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod common;
 pub mod clock;
 pub mod dom;
 pub mod proxy;
+pub mod telemetry;
